8:33 AM finally fixed the weird warning in PFParams; 9:28 AM fixed the cv_format misuses; 9:42 AM moved the input and pre proc creation code in runMTF to try blocks too; 1:17 PM fixed the remaining thirdparty warnings; 1:26 PM moved the remaining uses of the tracker object in runMTF within try catch blocks; 3:19 PM completed updating the readme; fixed the interface for estimateOpticalFlow; 3:26 PM changed n_pts to unsigned int in GridTrackerFlow; 3:47 PM made more improvements to the readme;

+++ Config/include/mtf/Config/parameters.h
@@ -55,7 +55,7 @@ namespace mtf{
 		/* default parameters */
 		int source_id = 0;
 		int actor_id = 0;
-		int n_trackers = 1;
+		unsigned int n_trackers = 1;
 		bool track_single_obj = false;
 		char pipeline = 'c';
 		char img_source = 'j';
@@ -977,7 +977,7 @@ namespace mtf{
 			char *arg_prefix = nullptr){
 			if(arg_val[0] == '#'){ return; }
 			if(!strcmp(arg_name, "n_trackers")){
-				n_trackers = atoi(arg_val);
+				n_trackers = static_cast<unsigned int>(atoi(arg_val));
 				return;
 			}
 			if(!strcmp(arg_name, "track_single_obj")){
